dislodetect_pybe/requirements.txt
Flask==3.0.0
flask_cors
gunicorn
Werkzeug==3.0.0
torch==2.3.1
torchvision==0.18.1
torchmetrics==1.4.0.post0
ultralytics==8.0.202
boto3
python-dotenv

